removed previous table and unused dnd component
