--- app/react/V2/Components/Layouts/DragAndDrop/SortFunction.ts
@@ -1,95 +1,0 @@
-import { RefObject } from 'react';
-import type { XYCoord } from 'react-dnd/dist/types/monitors';
-import { IDraggable } from 'app/V2/shared/types';
-
-const checkSortArea = (
-  monitor: any,
-  hoverBoundingRect: DOMRect,
-  dragIndex: number,
-  hoverIndex: number
-) => {
-  // Get vertical middle
-  const hoverMiddleY = (hoverBoundingRect.bottom - hoverBoundingRect.top) / 2;
-
-  // Determine mouse position
-  const clientOffset = monitor.getClientOffset();
-
-  // Get pixels to the top
-  const hoverClientY = (clientOffset as XYCoord).y - hoverBoundingRect.top;
-
-  // Only perform the move when the mouse has crossed half of the items height
-  // When dragging downwards, only move when the cursor is below 50%
-  // When dragging upwards, only move when the cursor is above 50%
-
-  // Dragging downwards OR Dragging upwards
-  if (
-    (dragIndex < hoverIndex && hoverClientY < hoverMiddleY) ||
-    (dragIndex > hoverIndex && hoverClientY > hoverMiddleY)
-  ) {
-    return true;
-  }
-  return false;
-};
-
-const isOutOfSortArea = (
-  ref: RefObject<HTMLElement>,
-  monitor: any,
-  dragIndex: number,
-  hoverIndex: number
-) =>
-  // Determine rectangle on screen
-  // No hoverBoundingRect AND Don't replace items with themselves
-  !ref.current ||
-  !ref.current.getBoundingClientRect() ||
-  dragIndex === hoverIndex ||
-  checkSortArea(monitor, ref.current.getBoundingClientRect(), dragIndex, hoverIndex);
-
-const isNotSortableItem = <T>(
-  currentItem: { dndId: string; item: IDraggable<T> },
-  target: IDraggable<T> & { ID?: string }
-) =>
-  (currentItem.item.parent && !target.parent) ||
-  currentItem.dndId === target.dndId ||
-  currentItem.item.container === undefined;
-
-const hoverSortable =
-  <T>(
-    ref: RefObject<HTMLElement>,
-    target: IDraggable<T>,
-    index: number,
-    setDragging: React.Dispatch<React.SetStateAction<boolean>>,
-    sortFunction?: Function
-  ) =>
-  (
-    currentItem: {
-      index: number;
-      dndId: string;
-      type: string;
-      item: IDraggable<T>;
-    },
-    monitor: any
-  ) => {
-    setDragging(true);
-    const dragIndex = currentItem.index;
-    const hoverIndex = index;
-    const draggingDown = monitor.getDifferenceFromInitialOffset().y > 0;
-    const invalidSort =
-      isNotSortableItem(currentItem, target) ||
-      (draggingDown && dragIndex > hoverIndex && hoverIndex === 0) ||
-      isOutOfSortArea(ref, monitor, dragIndex, hoverIndex);
-
-    if (!ref.current || !sortFunction || !monitor.isOver({ shallow: false }) || invalidSort) {
-      return;
-    }
-    // Time to actually perform the action
-    sortFunction(currentItem.item, target, dragIndex, hoverIndex);
-
-    // Note: we're mutating the monitor item here!
-    // Generally it's better to avoid mutations,
-    // but it's good here for the sake of performance
-    // to avoid expensive index searches.
-    // eslint-disable-next-line no-param-reassign
-    currentItem.index = hoverIndex;
-  };
-
-export { hoverSortable };
