--- app/react/V2/Components/Layouts/DragAndDrop/index.ts
@@ -1,7 +1,0 @@
-export { DropZone } from './DropZone';
-export type { IItemComponentProps } from './Container';
-export { Container } from './Container';
-export { DraggableItem } from './DraggableItem';
-export { DragSource } from './DragSource';
-export { useDnDContext } from './useDnDContext';
-export type { IDnDContext } from './DnDDefinitions';
