+++ app/react/V2/Components/UI/Table/index.ts
@@ -1,0 +1,2 @@
+export { Table } from './Table';
+export type { TableProps, TableRow } from './Table';
