+++ app/react/V2/Components/UI/index.ts
@@ -2,7 +2,7 @@ export { Button } from './Button';
 export { EmbededButton } from './EmbededButton';
 export { Modal } from './Modal';
 export { Pill } from './Pill';
-export { Table } from './TableV2/Table';
+export { Table } from './Table';
 export { ToggleButton } from './ToggleButton';
 export { NotificationsContainer } from './NotificationsContainer';
 export { Tabs } from './Tabs';
@@ -17,10 +17,4 @@ export { MediaPlayer } from './MediaPlayer';
 export { FileIcon } from './FileIcon';
 
 export type { PillColor } from './Pill';
-export type { TableProps, TableRow } from './TableV2/Table';
-
-//This component will be migrated to the new version of the table
-// eslint-disable-next-line camelcase
-export { Table as Table_deprecated } from './Table/Table';
-// eslint-disable-next-line camelcase
-export type { TableProps as TableProps_deprecated } from './Table/TableElements';
+export type { TableProps, TableRow } from './Table';
