+++ package.json
@@ -201,8 +201,6 @@
     "react-color": "^2.19.3",
     "react-datepicker": "7.3.0",
     "react-device-detect": "^2.2.3",
-    "react-dnd": "^16.0.1",
-    "react-dnd-html5-backend": "^16.0.1",
     "react-dnd-html5-backend-old": "yarn:react-dnd-html5-backend@^15.1.2",
     "react-dnd-old": "yarn:react-dnd@2.6.0",
     "react-dom": "^18.3.1",
